Fix failing Python tests

+++ pyvq/tests/test_pq.py
@@ -44,7 +44,7 @@ def test_product_quantizer_quantize():
     vector = training[0].copy()
     codes = pq.quantize(vector)
     assert isinstance(codes, np.ndarray)
-    assert codes.dtype == np.float32
+    assert codes.dtype == np.float16
     assert len(codes) == 12
